+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -103,6 +103,8 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 - **Breaking:** Consolidated WebView API options into a single object (`WebViewOpts` type), both for developer convenience and to support future options.
 - Enhanced the `ZoweVsCodeExtension.loginWithBaseProfile` and `ZoweVsCodeExtension.logoutWithBaseProfile` methods to store SSO token in parent profile when nested profiles are in use. [#2264](https://github.com/zowe/zowe-explorer-vscode/issues/2264)
 - **Next Breaking:** Changed return type of `ZoweVsCodeExtension.logoutWithBaseProfile` method from `void` to `boolean` to indicate whether logout was successful.
+- Changed `TableViewProvider.setTableView` function to be asynchronous for more optimized data updates.
+- Updated `Table.Conditional` and `Table.Callback` types to support multi-row callbacks.
 
 ### Bug fixes